refactor: extract response policy handling
